fixed bug with quitGame

+++ src/main/java/org/springframework/samples/parchisoca/service/GameService.java
@@ -180,8 +180,8 @@ public class GameService {
     {
         Optional<Game> gameOptional = this.findGamebyID(gameid);
         Game game = gameOptional.orElseThrow(EntityNotFoundException::new);
-
-
+        if(game.getStatus() == GameStatus.FINISHED)
+            return;
         game.setStatus(GameStatus.FINISHED);
         deleteAllGamePieces(game);
         game.setEndTime(LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS));
